+++ examples/SimpleExample.c
@@ -1,4 +1,5 @@
 #include <FrameLog.h>
+#include <stdio.h>
 
 int main(){
     FLPrintln("Hello, World!");
@@ -6,5 +7,7 @@ int main(){
     for(int i = 0; i < 10; i++){
         FLTraceln("output in loop!");
     }
+    int value = FLReadInt("");
+    printf("%i%s", value, "\n");
     FLFlushBuffer();
 }