+++ source/FrameLog/Prints.c
@@ -96,6 +96,82 @@ FL_API void FLErrorln(const char *Message)
     FLPrintColored(Message, FL_RED, NULL, true);
 }
 // =============================================================
+// Fatal Prints
+// =============================================================
+FL_API void FLFatal(const char* Message)
+{
+    FLPrintColored(Message, NULL, FL_REDB, false);
+}
+FL_API void FLFatalln(const char* Message){
+    FLPrintColored(Message, NULL, FL_REDB, true);
+}
+// =============================================================
+// Readers
+// =============================================================
+FL_API int FLReadInt(const char *Message){
+    int value = 0;
+    printf("%s", Message);
+
+
+    while (scanf("%d", &value) != 1) {
+        printf("FL Input Error: Input not number\n");
+        while (getchar() != '\n');
+    }
+    while (getchar() != '\n');
+    return value;
+}
+
+
+FL_API float FLReadFloat(const char* Message){
+    float value = 0;
+    printf("%s", Message);
+
+
+    while (scanf("%f", &value) != 1) {
+        printf("FL Input Error: Input not number\n");
+        while (getchar() != '\n');
+    }
+    while (getchar() != '\n');
+    return value;
+}
+
+
+FL_API double FLReadDouble(const char* Message){
+    double value = 0;
+    printf("%s", Message);
+
+
+    while (scanf("%lf", &value) != 1) {
+        printf("FL Input Error: Input not number\n");
+        while (getchar() != '\n');
+    }
+    while (getchar() != '\n');
+    return value;
+}
+
+
+FL_API char FLReadChar(const char* Message) {
+    char c = 0;
+    printf("%s", Message);
+
+    c = getchar();
+    while (getchar() != '\n');
+    return c;
+}
+
+
+FL_API const char* FLReadString(const char* Message) {
+    static char buffer[2000];
+    printf("%s", Message);
+
+
+    if (fgets(buffer, sizeof(buffer), stdin)) {
+        buffer[strcspn(buffer, "\n")] = '\0';
+        return buffer;
+    }
+    return "";
+}
+// =============================================================
 // Tools
 // =============================================================
 FL_API void FLNewLine()